% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_optional_dependencies}
\alias{check_optional_dependencies}
\title{Check GEDI Optional Dependencies}
\usage{
check_optional_dependencies()
}
\value{
Named logical vector indicating which optional packages are installed
}
\description{
Check which optional dependencies are installed and display their status.
}
\examples{
check_optional_dependencies()

}
