% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_plot_utils.R
\name{compute_color_limits}
\alias{compute_color_limits}
\title{Compute Color Limits from Data}
\usage{
compute_color_limits(values, symmetric = TRUE, quantile = 0.99)
}
\description{
Compute Color Limits from Data
}
\keyword{internal}
