% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_plot_wrappers.R
\name{.get_color_vector}
\alias{.get_color_vector}
\title{Get Color Vector from Model}
\usage{
.get_color_vector(model, color_by, projection = "zdb")
}
\arguments{
\item{model}{GEDI model object}

\item{color_by}{Character: "sample", metadata column name, or gene name/index}

\item{projection}{Character: "zdb" or "db" for gene expression}
}
\value{
Vector of length N with color values
}
\description{
Get Color Vector from Model
}
\keyword{internal}
