% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_dynamics.R
\name{print.gedi_dynamics}
\alias{print.gedi_dynamics}
\title{Print Method for Dynamics Accessor}
\usage{
print.gedi_dynamics(x, ...)
}
\arguments{
\item{x}{Object of class gedi_dynamics}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Dynamics Accessor
}
\keyword{internal}
