% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_dynamics.R
\name{print.gedi_dynamics_svd}
\alias{print.gedi_dynamics_svd}
\title{Print Method for Dynamics SVD Results}
\usage{
print.gedi_dynamics_svd(x, ...)
}
\arguments{
\item{x}{Object of class gedi_dynamics_svd}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Dynamics SVD Results
}
\keyword{internal}
