% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_accessors.R
\name{print.gedi_imputation}
\alias{print.gedi_imputation}
\title{Print Method for Imputation Accessor}
\usage{
print.gedi_imputation(x, ...)
}
\arguments{
\item{x}{Object of class gedi_imputation}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Imputation Accessor
}
\keyword{internal}
