% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_pathway.R
\name{print.gedi_pathway_associations}
\alias{print.gedi_pathway_associations}
\title{Print Method for Pathway Associations Accessor}
\usage{
print.gedi_pathway_associations(x, ...)
}
\arguments{
\item{x}{Object of class gedi_pathway_associations}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Pathway Associations Accessor
}
\keyword{internal}
