% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_plot_utils.R
\name{scale_color_gedi_discrete}
\alias{scale_color_gedi_discrete}
\title{Discrete Color Palette for GEDI}
\usage{
scale_color_gedi_discrete(name = "Group", ...)
}
\description{
Discrete Color Palette for GEDI
}
\keyword{internal}
