% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_plot_utils.R
\name{scale_color_gedi_diverging}
\alias{scale_color_gedi_diverging}
\title{Diverging Color Scale for GEDI Plots}
\usage{
scale_color_gedi_diverging(limits = NULL, name = "Value", ...)
}
\value{
A \code{ggplot2} continuous color scale.
}
\description{
Diverging Color Scale for GEDI Plots
}
\keyword{internal}
