% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedi_plot_utils.R
\name{scale_fill_gedi_diverging}
\alias{scale_fill_gedi_diverging}
\title{Fill Scale for GEDI Plots}
\usage{
scale_fill_gedi_diverging(limits = NULL, name = "Value", ...)
}
\description{
Fill Scale for GEDI Plots
}
\keyword{internal}
